+++ src/Blog/Post/Scope/PublicScope.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace App\Constraint;
+namespace App\Blog\Post\Scope;
 
 use Cycle\ORM\Select\ConstrainInterface;
 use Cycle\ORM\Select\QueryBuilder;
 
-class PostPublic implements ConstrainInterface
+class PublicScope implements ConstrainInterface
 {
     public function apply(QueryBuilder $query): void
     {
